requirements.txt
pytest
playwright
requests

# Зависимости для работы
pytest-playwright
pytest-html
pytest-xdist
allure-pytest

# Вспомогательные библиотеки
python-dotenv
pyyaml
jsonschema
faker
python-dotenv